# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/categories/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description" : "ok"
          }
        }
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/auth/sign-up": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/todos/create": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/todos/list": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/todos/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/todos/completed": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/todos/categories": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/todos/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    }
  }
}